requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
